Unified solution and integration of benchmark

+++ 2022/Day01/Item.cs
@@ -1,3 +1,3 @@
 ﻿namespace AdventOfCode.Day01;
 
-internal readonly record struct Item(int Index, int Value);
+internal readonly record struct Item(int Index, long Value);
